back mensageria validacoes 100%

+++ src/Application/UseCases/Repos/Remove/RemoveRepository.usecase.ts
@@ -1,6 +1,5 @@
 import { KEY_OF_INJECTION } from '@metadata';
 import { Inject, Injectable } from '@nestjs/common';
-import { RepositoryDTO } from 'src/@shared/@dtos';
 import { IRepoRepository } from 'src/Domain/Interfaces/Repositories/Repos/IRepo.repository';
 
 @Injectable()
@@ -11,6 +10,16 @@ export class DeleteRepository {
   ) {}
 
   async execute(repo: string) {
+    if (!repo || typeof repo !== 'string' || repo.trim() === '') {
+      throw new Error('Repository name is required and must be a non-empty string.');
+    }
+
+    const existingRepo = await this.repoRepository.getBy({ name: repo });
+
+    if (!existingRepo) {
+      throw new Error(`Cannot delete. Repository with name "${repo}" does not exist.`);
+    }
+
     await this.repoRepository.delete({ name: repo });
   }
 }
